requirements.txt
click
newsapi-python
arxiv
pandas
scikit-learn
sentence-transformers
beautifulsoup4
requests
streamlit
